FHAC-599 Addressed the PR comments
