+++ Product/Production/Services/PatientDiscoveryCore/src/main/java/gov/hhs/fha/nhinc/patientdiscovery/audit/transform/PatientDiscoveryAuditTransforms.java
@@ -27,15 +27,9 @@
 package gov.hhs.fha.nhinc.patientdiscovery.audit.transform;
 
 import com.services.nhinc.schema.auditmessage.AuditMessageType;
-import com.services.nhinc.schema.auditmessage.ParticipantObjectIdentificationType;
 import gov.hhs.fha.nhinc.common.nhinccommon.AssertionType;
-import gov.hhs.fha.nhinc.patientdiscovery.parser.PRPAIN201305UV02Parser;
-import gov.hhs.fha.nhinc.patientdiscovery.parser.PRPAIN201306UV02Parser;
-import java.io.ByteArrayOutputStream;
-import java.util.List;
 import javax.xml.bind.JAXBException;
 import org.apache.log4j.Logger;
-import org.hl7.v3.II;
 import org.hl7.v3.PRPAIN201305UV02;
 import org.hl7.v3.PRPAIN201306UV02;
 
@@ -78,57 +72,5 @@ public class PatientDiscoveryAuditTransforms extends AbstractPatientDiscoveryAud
         }
 
         return auditMsg;
-    }
-
-    private AuditMessageType getPatientParticipantObjectIdentificationForRequest(PRPAIN201305UV02 request,
-        AuditMessageType auditMsg) {
-
-        // Set the Partipation Object Id (patient id)
-        II oII = PRPAIN201305UV02Parser.getPatientId(request);
-        if (oII != null && oII.getRoot() != null && oII.getExtension() != null && !oII.getRoot().isEmpty()
-            && !oII.getExtension().isEmpty()) {
-
-            createPatientParticipantObjectIdentification(auditMsg, oII.getRoot(), oII.getExtension());
-        } else {
-            createPatientParticipantObjectIdentification(auditMsg, null, null);
-        }
-
-        return auditMsg;
-    }
-
-    private AuditMessageType getPatientParticipantObjectIdentificationForResponse(PRPAIN201306UV02 response,
-        AuditMessageType auditMsg) {
-
-        List<II> oII = PRPAIN201306UV02Parser.getPatientIds(response);
-        if (oII != null && oII.size() > 0) {
-            for (II entry : oII) {
-                if (entry != null && entry.getRoot() != null && entry.getExtension() != null
-                    && !entry.getRoot().isEmpty() && !entry.getExtension().isEmpty()) {
-                    createPatientParticipantObjectIdentification(auditMsg, entry.getRoot(), entry.getExtension());
-                }
-            }
-        } else {
-            createPatientParticipantObjectIdentification(auditMsg, null, null);
-        }
-
-        return auditMsg;
-    }
-
-    private AuditMessageType getQueryParamsParticipantObjectIdentificationForResponse(PRPAIN201306UV02 response,
-        AuditMessageType auditMsg) throws JAXBException {
-
-        ParticipantObjectIdentificationType participantObject = buildBaseParticipantObjectIdentificationType(
-            PRPAIN201306UV02Parser.getQueryId(response));
-        participantObject.setParticipantObjectQuery(getParticipantObjectQueryForResponse(response));
-        auditMsg.getParticipantObjectIdentification().add(participantObject);
-        return auditMsg;
-    }
-
-    private byte[] getParticipantObjectQueryForResponse(PRPAIN201306UV02 response) throws JAXBException {
-        ByteArrayOutputStream baos = new ByteArrayOutputStream();
-        if (response.getControlActProcess() != null && response.getControlActProcess().getQueryByParameter() != null) {
-            getMarshaller().marshal(response.getControlActProcess().getQueryByParameter(), baos);
-        }
-        return baos.toByteArray();
     }
 }
